:sparkles: feat: getZonePreset() helper
Returns the preset that's currently matching with the zone data as well as
with the devices part of the zone otherwise returns null.
This means if the zone has a `DeviceStatus.Comfort`, but only one of the
device from the zone has a different value, the function would return null.